Use one-time password for deployment
This commit introduces an additional environment variable which is
the one-time password generated by 2 factor authentication.

The Fastly gem doesn't support the tokens endpoint yet but once it
does we should refactor this to use it.

I've generated a token which is valid for 20 minutes and then use
that token for all the API requests.

This means that we can remove the user account we have which only
exists for deployment, and we can use our personal accounts which
have 2FA enabled. This is better for auditing who changed things
and for security.