Push to release branches after deploy
At the moment we don't know what's deployed to staging or production.
This commit adds a step to push to the deployed-to-staging and
deployed-to-production branches after each deploy. This will enable us
to compare the currently deployed version. It's consistent with what we
do in govuk-app-deployment[1]

[1]:
https://github.com/alphagov/govuk-app-deployment/blob/64941e7f0ca4b78ba6
bc745dc8dd9b369c6931d8/recipes/defaults.rb#L189