Don't delete Conditions created in the UI when deploying VCL
Condition objects are used to restrict which requests are logged for
specific log endpoints. They can’t be configured using VCL, so we
shouldn’t attempt to use the VCL as the source of truth.
Until we have Fastly configured using Terraform, we’ll need to carry on
setting up logging endpoints and related conditions in the UI.